pyproject.toml
[build-system]
requires = ["setuptools>=45", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "ragmatic"
version = "0.1.0"
description = "A package for using retrieval-augmented generation as a knowledge base for question-answering about a codebase."
authors = [{name = "Your Name", email = "your.email@example.com"}]
license = {text = "MIT"}
readme = "README.md"
requires-python = ">=3.10"
classifiers = [
    "Programming Language :: Python :: 3",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
]
dependencies = [
    "openai>=1.37.1",
    "tenacity>=9.0.0",
    "networkx>=3.3",
    "pydantic",
    "joblib>=1.4.2",
    "click>=8.1.7",
    "pyyaml>=6.0.1",
]

# define optional dependencies
[project.optional-dependencies]
qa = [
    "pytest",
    "pytest-cov",
    # "pytest-asyncio",
]

[project.scripts]
ragmatic = "ragmatic.cli:cli"
